Move R_CAllMethodDef back to inapplicable-init.c
